Version 0.5
I added support for for and while loops:  % and g% go between the start of the block and either break or continue lines.

I reorganized the code and made several minor improvements.  For example, [% and ]% now accept a count,  `` goes back to where you started, and g% does nothing (insteadof acting like the default %) if the cursor starts on a line that does not begin with a recognized key word.